new class, starting tests
